+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Change Log
 ### Improvements
 
 - Change the system requirement to __OS X 10.12.2 Sierra and later__.
+- [trivial] Increase contrast of the page guide in the high contrast mode.
 - [trivial] Optimize the performance of line number drawing.
 - [dev] Update build environment to Xcode 10 (Swift 4.2, macOS 10.14 SDK).